Shopp/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-colorfield
django-environ
django-tinymce
django-rosetta 
django-tinymce
djangorestframework
fontawesomefree
idna
localflavor
nltk
pandas
Pillow
psycopg2
pytz
requests
scipy
sqlparse
urllib3
weasyprint